+++ bright-ideas/src/app/services/idea-endpoint/idea-endpoint.service.ts
@@ -1,16 +1,15 @@
-import { Injectable } from '@angular/core';
-import { HttpClient } from '@angular/common/http';
-import { Observable } from 'rxjs';
-import { Idea } from '../models/idea'
+import { Injectable } from "@angular/core";
+import { HttpClient } from "@angular/common/http";
+import { Observable } from "rxjs";
+import { Idea } from "../../models/idea";
 
 @Injectable({
-  providedIn: 'root'
+  providedIn: "root"
 })
 export class IdeaEndpointService {
-
   private url = "https://bright-ideas-api.herokuapp.com";
 
-  constructor(private http: HttpClient) { }
+  constructor(private http: HttpClient) {}
 
   getIdeas(): Observable<Idea[]> {
     return this.http.get<Idea[]>(`${this.url}/ideas`);
