--- bright-ideas/src/app/services/todo-endpoint.service.spec.ts
@@ -1,12 +1,0 @@
-import { TestBed } from '@angular/core/testing';
-
-import { TodoEndpointService } from './todo-endpoint.service';
-
-describe('TodoEndpointService', () => {
-  beforeEach(() => TestBed.configureTestingModule({}));
-
-  it('should be created', () => {
-    const service: TodoEndpointService = TestBed.get(TodoEndpointService);
-    expect(service).toBeTruthy();
-  });
-});
