+++ bright-ideas/src/app/services/todo-endpoint/todo-endpoint.service.ts
@@ -1,17 +1,15 @@
-import { Injectable } from '@angular/core';
-import { HttpClient } from '@angular/common/http';
-import { Observable } from 'rxjs';
-import { ToDo } from '../models/todo'
+import { Injectable } from "@angular/core";
+import { HttpClient } from "@angular/common/http";
+import { Observable } from "rxjs";
+import { ToDo } from "../../models/todo";
 
 @Injectable({
-  providedIn: 'root'
+  providedIn: "root"
 })
-
 export class TodoEndpointService {
-
   private url = "https://bright-ideas-api.herokuapp.com";
 
-  constructor(private http: HttpClient) { }
+  constructor(private http: HttpClient) {}
 
   getAllToDo(): Observable<ToDo[]> {
     return this.http.get<ToDo[]>(`${this.url}/todos`);
