+++ bright-ideas/src/app/services/user-endpoint/user-endpoint.service.spec.ts
@@ -1,0 +1,12 @@
+import { TestBed } from "@angular/core/testing";
+
+import { UserEndpointService } from "../user-endpoint.service";
+
+describe("UserEndpointService", () => {
+  beforeEach(() => TestBed.configureTestingModule({}));
+
+  it("should be created", () => {
+    const service: UserEndpointService = TestBed.get(UserEndpointService);
+    expect(service).toBeTruthy();
+  });
+});
