+++ bright-ideas/src/app/services/user-endpoint/user-endpoint.service.ts
@@ -1,16 +1,15 @@
-import { Injectable } from '@angular/core';
-import { HttpClient } from '@angular/common/http';
-import { Observable } from 'rxjs';
-import { User } from '../models/user'
+import { Injectable } from "@angular/core";
+import { HttpClient } from "@angular/common/http";
+import { Observable } from "rxjs";
+import { User } from "../../models/user";
 
 @Injectable({
-  providedIn: 'root'
+  providedIn: "root"
 })
 export class UserEndpointService {
-
   private url = "https://bright-ideas-api.herokuapp.com";
 
-  constructor(private http: HttpClient) { }
+  constructor(private http: HttpClient) {}
 
   getUsers(): Observable<User[]> {
     return this.http.get<User[]>(`${this.url}/users`);
@@ -20,5 +19,3 @@ export class UserEndpointService {
     return this.http.get<User[]>(`${this.url}/user/` + id);
   }
 }
-
-
